changelog_9
- added ingame debug console plugin
- added dev menu for chat tests firstly
- added dev utility panel
utility panel for: add gold
- impoved spin handler also extra spin
- fixed bugs with onValidate in build
- impoved chat paddings
- fixed bug with incorrect index content in content screen

+++ Assets/_School_Seducer_/Editor/Scripts/Menu.cs
@@ -28,7 +28,7 @@ namespace _BonGirl_.Editor.Scripts
             Debug.Log("Panel selectedd: " + _selectedPanel.name);
         } 
 
-        public void EnableByChangePositionX(int xPosition)
+        public void EnableByChangePositionX(float xPosition)
         {
             if (_selectedPanel == null)
             {
